% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmfile.r
\name{bgmfile}
\alias{bgmfile}
\title{Read BGM}
\usage{
bgmfile(x, ...)
}
\arguments{
\item{x}{path to a bgm file}

\item{...}{ignored for now}
}
\description{
Partial read for .bgm files
}
\details{
Read geometry from BGM files
}

