% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbgm.R
\name{densifymindist}
\alias{densifymindist}
\title{Densify lonlat vertices}
\usage{
densifymindist(x, mindist, longlat = longlat)
}
\arguments{
\item{x}{matrix of 2 coordinates}

\item{mindist}{minimum distance in km if \code{longlat} is \code{TRUE}, otherwise in the units of the projection used (probably metres)}

\item{longlat}{calculate distance on great circle or Cartesian}
}
\value{
matrix of (possibly) densified coordinates
}
\description{
Densify vertices so that no span is  is greater than mindist
}
\details{
Using great circle distance in km add vertices to a matrix of longitude latitude coordinates
}

