% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgmfile.r
\name{rbgm-defunct}
\alias{rbgm-defunct}
\alias{read_bgm}
\title{Functions now defunct}
\usage{
read_bgm(x, ...)
}
\arguments{
\item{x}{path to a bgm file}

\item{...}{reserved}
}
\description{
read_bgm - use bgmfile
}

