% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbgm.R
\docType{package}
\name{rbgm-package}
\alias{rbgm-package}
\title{Utilities for BGM files for Atlantis}
\description{
Tools for handling network data for Atlantis from box-geometry
model (BGM) files
}
\author{
Michael D. Sumner
}

