% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbgm.R
\name{boxSpatial}
\alias{boxSpatial}
\alias{faceSpatial}
\title{Extract}
\usage{
boxSpatial(bgm)

faceSpatial(bgm)
}
\arguments{
\item{bgm}{}
}
\value{
Spatial* object 
\itemize{
\item boxSpatial \code{\link[sp]{SpatialPolygonsDataFrame}} 
\item faceSpatial  \code{\link[sp]{SpatialLinesDataFrame}} 
}
}
\description{
Take the output of \code{\link{read_bgm}} and return a \code{\link[sp]{Spatial}} object.
}
\examples{
fname <- system.file("extdata", "Antarctica_28", package = "rbgm")
bgm <- read_bgm(fname)
spdf <- boxSpatial(bgm)
sldf <- faceSpatial(bgm)

plot(boxSpatial(b), col = grey(seq(0, 1, length = nrow(b$boxes))))
plot(faceSpatial(b), col = rainbow(nrow(b$faces)), lwd = 2,  add = TRUE)
}

