% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbgm.R
\name{read.faces}
\alias{read.faces}
\title{Read BGM}
\usage{
read.faces(x, sp = TRUE, densify, ...)
}
\arguments{
\item{x}{bgm file}

\item{sp}{return Spatial object or list}

\item{densify}{factor to increase density of vertices}

\item{...}{arguments passed to \code{\link{gcIntermediate}}}
}
\value{
A SpatialLinesDataFrame if \code{sp} is \code{TRUE},
otherwise a list of 2x2 matrices containing the end point
coordinates, the list is a named vector with the label for each
face
}
\description{
Partial read for .bgm files
}
\details{
Read geometry from BGM files
}
\examples{
## pull in faces as a SpatialLinesDataFrame, densified to 30km between vertices
\dontrun{
x <- read.faces("file.bgm", densify = 30, sp = TRUE)
}
}

