% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.r
\name{nodeSpatial}
\alias{nodeSpatial}
\alias{pointSpatial}
\title{Obtain all vertices as a \code{\link[sp]{SpatialPointsDataFrame}}.}
\usage{
nodeSpatial(bgm)

pointSpatial(bgm)
}
\arguments{
\item{bgm}{BGM object from \code{\link{bgmfile}}}
}
\value{
\code{\link[sp]{SpatialPointsDataFrame}}
}
\description{
Obtain all vertices as a \code{\link[sp]{SpatialPointsDataFrame}}.
}

